kubeadm: use DefValue for the --kubeconfig flag
- Rename FindExistingKubeConfig to FindKubeConfigFile

- Cobra supports a DefValue option which can be used
to differentiate between the cases where the user set a flag
and when a flag was unset, while still adding a default value.
Use this in options/generic.go for the kubeconfig flag.

- Remove the FindKubeConfigFile() logic from `reset` and `upgrade`
as these are node level kubeadm commands.

- Default kubeconfig values to "" everywhere where FindKubeConfigFile
is used. This allows to search for existing kubeconfig locations.